+++ lib/src/app/view/app_state.dart
@@ -2,19 +2,10 @@ import 'package:counter_app/src/view.dart';
 
 // This is the 'View' of the application. The 'look and feel' of the app.
 class MyView extends AppState {
-  factory MyView() => _this ??= MyView._();
-  MyView._()
+  MyView()
       : super(
           debugShowCheckedModeBanner: false,
           switchUI: Prefs.getBool('switchUI'),
-          home: HomePage(),
+          inHome: () => HomePage(key: AppState.homeKey),
         );
-  static MyView _this;
-
-  @override
-  void reassemble() {
-    super.reassemble();
-    // Allow for 'hot reload'
-    _this = null;
-  }
 }